Add filter to accommodate multi-currency symbol
This will allow customers with multi-currency see the respective vendor
currency symbol in the PO template. This is made possible with the
addition of a custom twig filter "format_currency". Latest version of
twig already implements this filter but we are still on an older version
of twig at the moment